requirements.txt
numpy==2.2.4
pandas==2.2.3
pillow==11.2.1
streamlit==1.44.1